[refactor] use while statement insted of Iterator
