+++ java/src/libbun/parser/ssa/PHINode.java
@@ -1,21 +1,20 @@
 package libbun.parser.ssa;
 
-import java.util.ArrayList;
-
 import libbun.parser.ast.ZBlockNode;
 import libbun.parser.ast.ZLocalDefinedNode;
 import libbun.parser.ast.ZNode;
+import libbun.util.ZArray;
 
 
 public class PHINode extends ZLocalDefinedNode {
-	public ArrayList<ZNode> Args;
-	public ArrayList<ZBlockNode> Blocks;
+	public ZArray<ZNode> Args;
+	public ZArray<ZBlockNode> Blocks;
 	public ZNode Value;
 	public PHINode(ZNode Value) {
 		super(Value.ParentNode, Value.SourceToken, 0);
 		this.Value = Value;
-		this.Args = new ArrayList<ZNode>();
-		this.Blocks = new ArrayList<ZBlockNode>();
+		this.Args = new ZArray<ZNode>(new ZNode[0]);
+		this.Blocks = new ZArray<ZBlockNode>(new ZBlockNode[0]);
 	}
 	public void AddIncoming(ZBlockNode block, ZNode node) {
 		this.Blocks.add(block);
